requirements.txt
django
pyJWT
djangorestframework
requests
djangorestframework-simplejwt